Wait for producer to finish
